requirements.txt
requests==2.32.3
pandas==2.2.3
fpdf2==2.8.1
tabulate==0.9.0
certifi>=2024.7.4
openpyxl==3.1.5
